requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
peewee==3.17.0
pyaes==1.6.1
pyasn1==0.5.0
pydantic==1.10.13
pydantic-core==2.6.3
PyYAML==6.0.1
rsa==4.9
Telethon==1.31.1
typing-extensions==4.7.1
yarl==1.9.2
